actor/critic; response parsing of json and dicts; last message saving/overwriting

+++ session_model.py
@@ -8,6 +8,7 @@ import snakemd
 from snakemd import Inline, Paragraph, Raw
 
 from config import get_config
+from utils import get_valid_filename
 
 #
 # Models
@@ -43,6 +44,7 @@ class Session(BaseModel):
 
 
 def _mk_session_filename(name: str, dir: str) -> str:
+    name = get_valid_filename(name)
     session_file = os.path.join(dir, f"{name}.json")
     return session_file
 
@@ -66,7 +68,9 @@ def session_dir(session_dir: str) -> list[str]:
     sessions = []
     for file in os.listdir(session_dir):
         if file.endswith(".json"):
-            sessions.append(file.removesuffix(".json"))
+            session = Session.parse_file(os.path.join(session_dir, file))
+            sessions.append(session.name)
+    sessions.sort()
     return sessions
 
 
@@ -74,10 +78,9 @@ def session_purge(session_dir: str, exclude: Session) -> None:
     for file in os.listdir(session_dir):
         if file.endswith(".json"):
             filename = file.removesuffix(".json")
-            if filename != exclude.name:
-                session = Session.parse_file(os.path.join(session_dir, file))
-                if len(session.messages) < 2:
-                    os.remove(os.path.join(session_dir, file))
+            session = Session.parse_file(os.path.join(session_dir, file))
+            if session.name != exclude.name and len(session.messages) < 2:
+                os.remove(os.path.join(session_dir, file))
 
 
 def session_load(name: str, session_dir: str) -> Session:
@@ -93,6 +96,7 @@ def _mk_transcript_filename(
 ) -> str:
     # ensure outdir exists
     os.makedirs(dir, exist_ok=True)
+    name = get_valid_filename(name)
     transcript_file = os.path.join(dir, f"{name} {timestamp}")
     return transcript_file
 
@@ -116,7 +120,7 @@ def session_transcript_save(
     doc = snakemd.new_doc()
     doc.add_heading(session.name, 1)
     doc.add_paragraph(
-        f"A Red v Blue transcript between {session.red_bot} (red) and {session.blue_bot} (blue) on {timestamp.strftime('%B %d, %Y at %I:%M %p')}."
+        f"A InsightCrafter™ transcript between {session.red_bot} (red) and {session.blue_bot} (blue) on {timestamp.strftime('%B %d, %Y at %I:%M %p')}."
     )
 
     doc.add_heading(f"Directives", 2)
